pyproject.toml
[tool.poetry]
name = "vision-test"
version = "0.1.0"
description = ""
authors = ["Binh Nguyen <b.nguyen.sn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.3"
opencv-python = "4.8.0.76"
ffmpeg-python = "0.2.0"
numpy = "1.25.2"
future = "0.18.3"
functions-framework = "3.8.1"
pyannote-audio = "3.3.1"
python-dotenv = "1.0.1"
huggingface-hub = "^0.24.6"
torch = {version = "^2.4.0+cu124", source = "pytorch-gpu-src"}
torchvision = {version = "^0.19.0+cu124", source = "pytorch-gpu-src"}
torchaudio = {version = "^2.4.0+cu124", source = "pytorch-gpu-src"}
psutil = "6.0.0"

[tool.poetry.scripts]
run-vision-test = "runner:main"
analyze-audio = "analyze_audio:main"
split-wav = "split_wav:main"
cv-playground = "cv_playground:main"

[[tool.poetry.source]]
name = "pytorch-gpu-src"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
